rename ZeUI package and zeclasses
