+++ zeapp/app/src/main/java/de/berlindroid/zeapp/MainActivity.kt
@@ -58,18 +58,18 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.tooling.preview.PreviewParameter
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import de.berlindroid.zeapp.ui.BinaryBitmapPageProvider
-import de.berlindroid.zeapp.ui.ImageGenerationEditorDialog
-import de.berlindroid.zeapp.ui.NameEditorDialog
-import de.berlindroid.zeapp.ui.PictureEditorDialog
-import de.berlindroid.zeapp.ui.QRCodeEditorDialog
-import de.berlindroid.zeapp.ui.ToolButton as ZeToolButton
-import de.berlindroid.zeapp.ui.zetheme.ZeBadgeAppTheme
+import de.berlindroid.zeapp.zeui.BinaryBitmapPageProvider
+import de.berlindroid.zeapp.zeui.ImageGenerationEditorDialog
+import de.berlindroid.zeapp.zeui.NameEditorDialog
+import de.berlindroid.zeapp.zeui.PictureEditorDialog
+import de.berlindroid.zeapp.zeui.QRCodeEditorDialog
+import de.berlindroid.zeapp.zeui.ToolButton as ZeToolButton
+import de.berlindroid.zeapp.zeui.zetheme.ZeBadgeAppTheme
 import de.berlindroid.zeapp.vm.ZeBadgeViewModel
 import de.berlindroid.zeapp.vm.ZeBadgeViewModel.*
 import android.content.res.Configuration as AndroidConfig
 import androidx.compose.material3.IconButton as ZeIconButton
-import de.berlindroid.zeapp.ui.BadgeSimulator as ZeSimulator
+import de.berlindroid.zeapp.zeui.BadgeSimulator as ZeSimulator
 
 /**
  * Main View entrance for the app
