+++ zeapp/app/src/main/java/de/berlindroid/zeapp/zeui/ZePictureEditorDialog.kt
@@ -1,4 +1,4 @@
-package de.berlindroid.zeapp.ui
+package de.berlindroid.zeapp.zeui
 
 import android.graphics.BitmapFactory
 import android.util.Log
