+++ zeapp/app/src/main/java/de/berlindroid/zeapp/zeui/ZeQRCodeEditorDialog.kt
@@ -1,6 +1,6 @@
 @file:OptIn(ExperimentalMaterial3Api::class, ExperimentalMaterial3Api::class)
 
-package de.berlindroid.zeapp.ui
+package de.berlindroid.zeapp.zeui
 
 import android.R
 import android.app.Activity
@@ -10,8 +10,8 @@ import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.material3.AlertDialog
 import androidx.compose.material3.Button
 import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.OutlinedTextField
 import androidx.compose.material3.Text
+import androidx.compose.material3.TextField
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
@@ -20,11 +20,9 @@ import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.window.DialogProperties
-import de.berlindroid.zeapp.zebits.composableToBitmap
 import de.berlindroid.zeapp.zebits.isBinary
-import de.berlindroid.zeapp.ui.zepages.NamePage
+import de.berlindroid.zeapp.zebits.qrComposableToBitmap
 import de.berlindroid.zeapp.vm.ZeBadgeViewModel.Configuration
-
 
 /**
  * Editor dialog for changing the name of the participant badge.
@@ -34,24 +32,22 @@ import de.berlindroid.zeapp.vm.ZeBadgeViewModel.Configuration
  * @param dismissed callback called when dialog is dismissed / cancelled
  * @param accepted callback called with the new configuration configured.
  */
-
-const val MaxCharacters: Int = 20
-
 @Composable
-fun NameEditorDialog(
+fun QRCodeEditorDialog(
     activity: Activity,
-    config: Configuration.Name,
+    config: Configuration.QRCode,
     dismissed: () -> Unit = {},
-    accepted: (config: Configuration.Name) -> Unit
+    accepted: (config: Configuration.QRCode) -> Unit
 ) {
-    var name by remember { mutableStateOf(config.name) }
-    var contact by remember { mutableStateOf(config.contact) }
+    var title by remember { mutableStateOf(config.title) }
+    var url by remember { mutableStateOf(config.url) }
     var image by remember { mutableStateOf(config.bitmap) }
 
     fun redrawComposableImage() {
-        composableToBitmap(
+        qrComposableToBitmap(
             activity = activity,
-            content = { NamePage(name, contact) },
+            title = title,
+            url = url,
         ) {
             image = it
         }
@@ -63,7 +59,7 @@ fun NameEditorDialog(
             Button(
                 onClick = {
                     if (image.isBinary()) {
-                        accepted(Configuration.Name(name, contact, image))
+                        accepted(Configuration.QRCode(title, url, image))
                     } else {
                         Toast.makeText(
                             activity,
@@ -75,12 +71,7 @@ fun NameEditorDialog(
                 Text(text = stringResource(id = R.string.ok))
             }
         },
-        dismissButton = {
-            Button(onClick = dismissed) {
-                Text(text = "Cancel")
-            }
-        },
-        title = { Text(text = "Add your contact details") },
+        title = { Text(text = "Add your QR url") },
         properties = DialogProperties(),
         text = {
             LazyColumn {
@@ -92,38 +83,27 @@ fun NameEditorDialog(
                 }
 
                 item {
-                    OutlinedTextField(
+                    TextField(
                         modifier = Modifier.fillMaxWidth(),
-                        value = name,
+                        value = title,
                         maxLines = 1,
-                        label = { Text(text = "Name") },
+                        label = { Text(text = "QR Code title") },
                         onValueChange = { newValue ->
-                            if (newValue.length <= MaxCharacters * 2) {
-                                name = newValue
-                                redrawComposableImage()
-                            }
-                        },
-                        supportingText = {
-                            Text(text = "${name.length}/${MaxCharacters * 2}")
+                            title = newValue
+                            redrawComposableImage()
                         }
                     )
                 }
 
                 item {
-                    OutlinedTextField(
+                    TextField(
                         modifier = Modifier.fillMaxWidth(),
-                        value = contact,
+                        value = url,
                         maxLines = 1,
-                        label = { Text(text = "Contact") },
+                        label = { Text(text = "URL") },
                         onValueChange = { newValue ->
-                            // Limit Characters so they're displayed correctly in the screen
-                            if (newValue.length <= MaxCharacters) {
-                                contact = newValue
-                                redrawComposableImage()
-                            }
-                        },
-                        supportingText = {
-                            Text(text = "${contact.length}/$MaxCharacters")
+                            url = newValue
+                            redrawComposableImage()
                         }
                     )
                 }
