+++ zeapp/app/src/main/java/de/berlindroid/zeapp/zeui/zepages/ZeQRCodePage.kt
@@ -1,38 +1,37 @@
-package de.berlindroid.zeapp.ui.zepages
+package de.berlindroid.zeapp.zeui.zepages
 
+import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.wrapContentSize
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.platform.LocalDensity
+import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.text.font.FontFamily
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import de.berlindroid.zeapp.PAGE_HEIGHT
 import de.berlindroid.zeapp.PAGE_WIDTH
+import de.berlindroid.zeapp.R
+import de.berlindroid.zeapp.zeui.rememberQrBitmapPainter
 
 /**
- * A composable that represents a badge, to be renderd as a bitmap.
+ * A composable that represents a badge, to be rendered as a bitmap.
  */
 @Composable
 @Preview
-fun NamePage(
-    name: String = "Your Name Here",
-    contact: String = "Contact Me Here",
+fun QRCodePage(
+    title: String = "Hello, my github profile is",
+    url: String = "https://github.com/gdg-berlin-android",
 ) {
-    var size by remember { mutableStateOf(10.sp) }
-
     Column(
         modifier = Modifier
             .background(
@@ -53,31 +52,17 @@ fun NamePage(
             textAlign = TextAlign.Center,
             color = Color.White,
             maxLines = 1,
-            text = "Hello, my name is",
+            text = title,
         )
-        Text(
-            modifier = Modifier
-                .wrapContentSize()
-                .weight(1.0f),
-            fontSize = size,
-            textAlign = TextAlign.Center,
-            text = name,
-            onTextLayout = { result ->
-                if (result.didOverflowWidth) {
-                    size /= 2.0f
-                }
-            },
-        )
-        Text(
-            modifier = Modifier
-                .fillMaxWidth()
-                .background(Color(0xFF801000)),
-            fontFamily = FontFamily.Monospace,
-            fontSize = 8.sp,
-            color = Color.White,
-            textAlign = TextAlign.Center,
-            maxLines = 1,
-            text = contact,
+
+        val painter =
+            if (url.isEmpty()) painterResource(id = R.drawable.page_telekom)
+            else rememberQrBitmapPainter(url)
+        Image(
+            painter = painter,
+            contentDescription = "",
+            contentScale = ContentScale.Fit,
+            modifier = Modifier.size(40.dp),
         )
     }
 }
