+++ zeapp/app/src/main/java/de/berlindroid/zeapp/zeui/zetheme/ZeColor.kt
@@ -1,4 +1,4 @@
-package de.berlindroid.zeapp.ui.zetheme
+package de.berlindroid.zeapp.zeui.zetheme
 
 import androidx.compose.ui.graphics.Color
