+++ zeapp/app/src/main/java/de/berlindroid/zeapp/zeui/zetheme/ZeTheme.kt
@@ -1,4 +1,4 @@
-package de.berlindroid.zeapp.ui.zetheme
+package de.berlindroid.zeapp.zeui.zetheme
 
 import android.app.Activity
 import android.os.Build
